fix: move topLevelGroupRules to its own module
