+++ core/src/actions/groups.ts
@@ -1,7 +1,8 @@
 import { CLS } from "../modules/cls";
 import { AuthenticatedAction } from "../classes/actions/authenticatedAction";
-import { Group, GROUP_RULE_LIMIT, TopLevelGroupRules } from "../models/Group";
+import { Group, GROUP_RULE_LIMIT } from "../models/Group";
 import { PropertyOpsDictionary } from "../modules/ruleOpsDictionary";
+import { TopLevelGroupRules } from "../modules/topLevelGroupRules";
 import { Profile } from "../models/Profile";
 import { GroupMember } from "../models/GroupMember";
 import { ConfigWriter } from "../modules/configWriter";
