+++ core/src/modules/topLevelGroupRules.ts
@@ -1,0 +1,4 @@
+export const TopLevelGroupRules = [
+  { key: "grouparooId", column: "id", type: "string" },
+  { key: "grouparooCreatedAt", column: "createdAt", type: "date" },
+];
